Major Bug Fix: Problem for training MultiModel classes - READ DESCRIPTION
Was overwriting results from previous folds when running cross-validation, leading to incorrect statistics when computing the performance statistics. 

Occurred because split models were computing training statistics for the sub models, which overwrites the predicted class variable for each entry. So, for folds 1-(k-1) during cross-validation, the predicted class variable from when that fold was used as the test fold was overwritten when those folds were used as the training set for the last fold. So, the CV results contained the average of the training errors for folds 1-(k-1) and the validation results for k fold.