Removed PerceptronRegression - Less featured than WekaRegression + functions.MultilayerPerceptron
You actually can have greater control over the NN using the WekaRegression interface. Removed to encourage users to use that for the time being

Eventually, I want to employ a tool such as Theano to run Deep NNs. For that, I'll likely use a similar approach to ScikitLearnRegression, where you supply an input file to Magpie and the model is run in separate thread.