Bug fix: Problems with the forName operation from Weka
I don't fully understand the error. Evidence

1. Scripts involving Weka run fine from Netbeans
2. Tests run fine on Netbeans and via Gradle
3. Build via Gradle works fine (Gradle makes a single jar with all of the dependancies inside)
4. Build with Ant doesn't work I get a "Can't find class" error from Weka when creating a model

Ant is the only of these options (I think) that involves putting class path in the jar manifest. Perhaps this error therefore has something to do with adding external jars to the classpath via the manifest. But, I don't understand the class loader enough to figure out why that's important.

This did work fine with Weka 3.7.9. Maybe there was a change between 9 - 13 in the Weka class loader.