Refactored how oxidation states are  handled
If the attributes are computed in parallel for a CompostionDataset, then the original object (i.e., the one that persists after attribute calculation) does not recieve a copy of the OxidationStates. All of the threads get their own copy of the primitive array and, therefore, have different values for this.OxidationStates. And, as the original dataset never loads in the oxidation states, its value remains unchanged "null"

In contrast, ElementalProperties Map pointed to in the original class is the same one pointed to by each thread (adding new properties does not change the address of the Map). Consequently, the original Dataset object points to the same Map as each of the threads and, therefore, has all of the same lookup data. 

In order to ensure a copy of the OxidationStates table is stored with the original class if it is saved during serialization, we check for the presence of OxidationStates in the LookupData object (being populated if another thread requests that data) and store it with this class