Also use attributes when comparing CompositionEntries
In most cases, entries with the same composition will
have the same attributes. However, if one uses a property
(maybe a processing condition) as an attribute, that
makes the entries different in terms of how they are
treated by the model. Magpie will now treat entries
with the same composition and different attributes
as different